feat(comp-face): Stage B — server-render the initial competition state
Convert the Live face route to a Server Component that resolves the
competition face's single boundary resolve (competitions.faceBootstrap) on
the server and hands it to the client face as initialData. The board/guide
now render populated in the server HTML — zero client round-trip for first
paint — and the client reads the data as fresh under the 60s staleTime, so
there is no mount refetch (one resolve per load).

- page.tsx becomes an async Server Component: createSSRHelpers().competitions
  .faceBootstrap.fetch({tripId}), passed down as initialBoot. Resolve failures
  are swallowed (mirrors the trip layout) → the client falls back to its own
  fetch + loading state, so an unauthed/early request never trips the route
  error boundary.
- New LiveFaceClient ("use client") holds all interactivity + realtime — the
  toggle, go-live/edit controls, chat/news, useRealtime* subscriptions — over
  the server-rendered initial state. It seeds the child caches from the same
  bootstrap during the SSR render too, so the whole face (board AND setup
  guide) ships populated. It is the SAME resolver as Stage A, just called
  server-side (B2); the seam between server data and client interactivity is
  this prop (B1). Shape is realtime-ready (B3) without building realtime here.

Verified live (owner path): the board is in the initial SSR HTML (no spinner
in <main>), there is NO client faceBootstrap round-trip (one resolve per load,
B4), the setup<->leaderboard toggle works post-hydration with no spinner, and
there are no hydration/console errors. tsc + lint clean. Role is still
live-derived per request server-side (faceBootstrap.test.ts covers
owner/co_admin/member/delegate).
